requirements.txt
flask
python-dotenv
google-generativeai
requests
requests-oauthlib
psycopg2-binary
gunicorn